Document onRender accessibility requirement (#1560)
For keyboard accessibility, all custom-rendered context menus entries need to be
focusable so that the FocusZone can set the appropriate tabIndex while
navigating through the menu.

Previously, this requirement was not documented which made it very easy to
inadvertently create non-accessible context menus. For example, the
"ContextualMenu customization example" on
https://dev.office.com/fabric#/components/contextualmenu is currently not
keyboard-accessible.