DetailsList: resize columns in justified mode doesn't affect columns on the left (#2995)
* Adjust only right-bound columns in justified mode.

Previously, columns on the left could change their sizes proportionally, which resulted on a weird experience for the final user.

* Fix fixedWidth formula so right-most column drops gracefully.

* Last column takes it all.

* Try to shrink last column before dropping it.

* Last column should be resizable if there is a dropped column.

* Non-collapsible columns should stay min when overflow happens (because we cannot drop them).
Also resizing last column should never drop it (we were doing > instead of >=).

* Missing semicolon (thanks TSLint).

* Extract method for readability, added summary description.

* Add change request.

* Allow last column to be resizable in justified layout.

This lets the last column to be extended further the right edge, and aligns with the idea that we should never prevent users to make any column wider if they need so. Note this will require a horizontal scrollbar when column overflows.

* Prevent non-collapsible columns to get smaller than its minWidth.

This happened when a previous column was resized till the right edge, the other columns became 0 (only visible because of padding).