Panel: pass ID for header text element to custom renderer (#4469)
The dialog inside Panel uses the panel header text as its ARIA label, referencing the header text element using `aria-labelledby=<header text id>`. That works fine with the default renderer, but custom renderers had no way to assign the correct ID, meaning the panel title was not read.

As a fix, this PR adds an additional `headerTextId` param to the header renderer.