Make examples using exampleData work in codepen + codepen-ify more examples (#7648)
Component examples which used office-ui-fabric-react/src/utilities/exampleData wouldn't work in codepen since the exampleData code is not included in the office-ui-fabric-react bundle. This change updates the codepen generation code to detect if a file imports exampleData, and if so, append the entire contents of the exampleData file to the end of the codepen file.