Fix ColorPicker when alpha slider is hidden (#7785)
When the ColorPicker's alpha slider was hidden, for some reason it wouldn't respect updates to the text fields. It seems this had something to do with the odd way that the ColorPicker was using text fields, because when the text field usage was switched to a more idiomatic controlled pattern, the problem went away.