Force Suggestions to be shrinkable (#7914)
* Force Suggestions to be shrinkable

Flexbox won't shrink an item below its intrinsic width. For nowrap text
(in a persona, for example), this means that when the item button
was rendered next to the remove button, it pushed the remove button out
of view.

This fixes this problem (#7912) by forcing the itemButton to be
shrinkable (min-width:0 ).

* add changefile

* revert autoformat

* remove unrelated changes

* Update packages/office-ui-fabric-react/src/components/pickers/Suggestions/Suggestions.scss

Co-Authored-By: Adjective-Object <mhuan13@gmail.com>

* Update packages/office-ui-fabric-react/src/components/pickers/Suggestions/SuggestionsItem.styles.ts

Co-Authored-By: Adjective-Object <mhuan13@gmail.com>

* Update packages/office-ui-fabric-react/src/components/pickers/Suggestions/Suggestions.scss

Co-Authored-By: Adjective-Object <mhuan13@gmail.com>

* Update packages/office-ui-fabric-react/src/components/pickers/Suggestions/SuggestionsItem.styles.ts

Co-Authored-By: Adjective-Object <mhuan13@gmail.com>

* suggestions snap

* add story for screener

* fix the screener story

* rewrite story to be of Provinces of Canada

* fixup test selector

* text alignment / overflow fix for webkit

* Mark story as dev-only

Mark the story as dev-only in the storybook and add a visual header to
the story to mark it as protecting a specific rendering edge-case, and
not the visual design itself.