Modal ignoreExternalFocusing prop should also affect Popup (#11535)
#### Pull request checklist

- [x] Addresses an existing issue: Fixes #11514
- [x] Include a change request file using `$ yarn change`

#### Description of changes

Modal/Dialog have a (misleadingly named) `ignoreExternalFocusing` prop which if true means that when the modal closes, it shouldn't re-focus the element which opened the modal. Popup (underlying component) *also* has the capability to re-focus elements, controlled by its `shouldRestoreFocus` prop. #11514 pointed out that even if re-focusing the opening element was disabled at the Modal level, that wasn't being passed through to the Popup--now fixed.

I also updated the docs for various `IAccessiblePopupProps` (Modal/Dialog shared) and `IFocusTrapZoneProps` members to be clearer/more accurate, and gave `IAccessiblePopupProps` a `docCategory` to make it actually show up on the website.