[BaseFloatingPicker] Pass custom props to inner Callout (#11794)
#### Pull request checklist

- [ ] Addresses an existing issue: Fixes #0000
- [x] Include a change request file using `$ yarn change`

#### Description of changes

There currently is a discrepancy between `BaseFloatingPicker` and `BasePicker` that the latter allows consumer to pass a `pickerCalloutProps` while the former doesn't thus preventing customization of the inner `Callout` component.

This PR fixes the issue by adding a `pickerCalloutProps` to `BaseFloatingPicker`'s props, and by doing so match `BasePicker`'s behavior.

#### Focus areas to test

This should be a fully backward-compatible fix. Currently behavior of `BaseFloatingPicker` should not be impacted.