Adding onClick callback to divider in SplitButtons that stops propagation so that clicking on the divider does not trigger action (#12325)
#### Pull request checklist

- [x] Addresses an existing issue: Fixes #12307
- [x] Include a change request file using `$ yarn change`

#### Description of changes

There was an issue where hovering over the divider in a `SplitButton` was indicating that clicking on it would not trigger anything but actually triggered the main action of the button. This was happening because the event was propagating to the button container. This PR adds a click callback to the divider to stop this propagation and make clicking on the divider actually reflect the affordance that the hover styling is showing the users.

#### Focus areas to test

(optional)