chore(FocusZone): Using the same DOM utilities in v0 that we use in v7 (#12192)
#### Pull request checklist

- [ ] Addresses an existing issue: Fixes #0000
- [x] Include a change request file using `$ yarn change`

#### Description of changes

This PR is part of a series of changes to bring the `FocusZone` in v0 and v7 closer together. This PR specifically deals with replacing v0's DOM utilities used in the `FocusZone` component with those that are used in v7. That means, for the moment, using utilities in the `uifabric/utilities` package. The location of these utilities will change in the future to a more appropriate place after the renaming has taken place, however, it is good to start using the same set of utilities so that things move more smoothly once the utilities breakup work starts.

In order to use `uifabric/utilities` inside v0's `FocusZone` we had to make some changes inside `uifabric/utilities` and `uifabric/merge-styles`. These changes are also included as part of this PR.

#### Focus areas to test

(optional)