ColorPicker: Add support for selecting transparency instead of alpha (#12153)
#### Pull request checklist

- [x] Include a change request file using `$ yarn change`

#### Description of changes

Add support for selecting transparency values instead of alpha values in the ColorPicker. This was started by @navkuma in #11836, but after thinking a good bit about what the best approach would be, I decided to make a separate PR implementing what I was thinking of.

The feature is enabled by setting a new prop `alphaType: 'alpha' | 'transparency' | 'none'`. This prop supersedes `alphaSliderHidden` and provides a single place to control if/how alpha/transparency controls are displayed. (Also considered `useTransparency` or `useTransparencySlider`, but `alphaType` has the advantage of putting all control in one place and mirroring the new `IColorSliderProps.type`. Of the other two options, I prefer `useTransparency` because the setting affects a text field, not just the slider.)

This is what it looks like in transparency mode:
![image](https://user-images.githubusercontent.com/5864305/75604281-d775d880-5a8b-11ea-856b-f96ab6c47359.png)

vs. alpha mode:
![image](https://user-images.githubusercontent.com/5864305/75604277-ce850700-5a8b-11ea-97a1-2b9a9d7ef23a.png)

Transparency values are stored in a new `IColor` property `t` (for clarity to the consumer). I updated the color utility functions ensure this mirrors the alpha value `IColor.a`.

For `ColorSlider`, to handle transparency sliders, I deprecated the `isAlpha` prop and added a new prop `type: 'hue' | 'alpha' | 'transparency'` (and deprecated `minValue` and `maxValue`, which are inferred from `type`). This is cleaner than having multiple boolean props for slider types.