Re-enable editor with better error handling (#14078)
#### Pull request checklist

- [x] Addresses an existing issue: Fixes #12754
- [x] Include a change request file using `$ yarn change`

#### Description of changes

#12776 disabled the editor for all examples since the editor was causing out of memory crashes for the list examples (see #12754) and we couldn't quickly figure out why. We think the issue in the list examples has to do with measurement, but it doesn't repro in local builds and therefore is very hard to debug, and no one has had time to look into it more deeply. (I also can't repro it now on the live site when I enable the editor with `?useEditor=1`.)

Instead of leaving the editor disabled indefinitely, we should re-enable it with additional logic to hopefully prevent infinite rendering loops in error cases. Generally the editor tries to re-render the last known good children when there's an error, but this was probably contributing to the infinite loop issue. So disable that behavior ~~entirely for list-related examples, and also disable it~~ for any example if more than one error is caught in < 1 second. (EDIT: removed list special case)