Remove lists and migration packages (#14994)
#### Pull request checklist

- [x] Addresses an existing issue: Somewhat related to #13384 
- [x] Include a change request file using `$ yarn change`

#### Description of changes

Delete the `@uifabric/lists` and `@uifabric/migration` packages. 

Per @soettl the lists package will not be shipped in its current form, and a version is in development internally which will eventually replace it, so there's really no point in keeping the code around. 

The migration package was used for v7 but has been replaced with the codemods package going forward.