Upload demo images to CDN (#15097)
Upload images used in screener and other places to our CDN that we use for the website (http://fabricweb.azureedge.net/fabric-website/*). This will make them load faster than if we uploaded them to normal blob storages, which can be important for getting correct results in screener tests.

For testing purposes I have the images uploading as part of the PR build, but before check-in I plan to change it so the images upload **only on master builds** (mainly to avoid accidental changes). This means that if you want to use a new image, you'll have to check it in separately first, but that shouldn't be a big issue since it won't happen very often.

Also update other upload tasks to use the new AzureUpload task which gzips things. This should make the PR deploy site load faster.