Button: Adding transparent variant (#15463)
<!--
!!!!!!! IMPORTANT !!!!!!!

Due to work we're currently doing to prepare master branch for our version 8 beta release,
please hold-off submitting the PR until around October 12 if it's not urgent.
If it is urgent, please submit the PR targeting the 7.0 branch.

This change does not apply to react-northstar contributors.

See https://github.com/microsoft/fluentui/issues/15222 for more details. Sorry for the inconvenience and short notice.
-->

#### Pull request checklist

- [x] Addresses an existing issue: Fixes #15105
- [x] Include a change request file using `$ yarn change`

#### Description of changes

This PR adds the `transparent` variant button that imitates the styling of the `ActionButton` in v7. Examples and visual regression tests have been updated to reflect this.

#### Focus areas to test

(optional)