fix: Improve variant style inheritance in Text component (#15942)
#### Pull request checklist

- [x] Include a change request file using `$ yarn change`

#### Description of changes

When applying custom styles to a font variant using `loadTheme()`, majority of the custom styles are not persisted to the style settings of the Text component.

The component has previously taken *all style settings* of the `medium` variant as base, however it would only apply a handful of style properties from the style settings of the actual variant. The pull request fixes this issue by taking the actual variant of the component as base while maintaining `medium` as a fallback option.