Cherry pick tooltip propagation fix (#16011)
#### Pull request checklist

- [ ] Addresses an existing issue: Fixes #0000
- [X] Include a change request file using `$ yarn change`

#### Description of changes

Currently, we are attempting to capture keydown events on the tooltip (particularly the Escape and Ctrl key) even when the tooltip is closed. This means that in the event that the tooltip belongs to an element in a dialog or another component that listens to the same keydown event, the event is swallowed. We should check to see that the tooltip is open before attempting to handle this event (so that we don't stop event propagation).

This is a cherry-pick from a change that was recently merged into v7.0 in this pull request.
https://github.com/microsoft/fluentui/pull/15990

#### Focus areas to test

(optional)