refactor(Tree): BREAKING CHANGE - rework tree as hook (#15831)
#### Pull request checklist

- [ ] Addresses an existing issue: Fixes #0000
- [ ] Include a change request file using `$ yarn change`

#### Description of changes

## Description of the re-work ideas:

The key is to create a **flattened tree** (see `flattenTree.ts`).
The flattened tree is an object, with key being the id of each tree item, and value being the information of the corresponding tree item: its parent/child, was it selected, ..., etc.

Initially the flattened tree is created thru Tree's `props.items`.
Then it gets re-created every time when state changes. Meaning every time users select/unselect, expand/collapse a tree node, there will be O(n) time spent on re-creating flattened tree, where n being number of tree items.

Why flatten tree:
- provides a convenient way of navigating up/down tree. 
- being the single source of truth for the state (expanded/collapsed, selected) of the tree
- together with `getItemById`, provides a context for tree items to access tree information.

The flattened tree is then used in `useTree` hook. This hook is a building block for `Tree` component, and a `useVirtualTree` hook (not in this PR yet) for virtualized tree component.
 
## Breaking changes:

- The following props are removed from `TreeItem`:

  ```ts
    /** A state of selection indicator. */
    selected?: boolean;

    /** Whether or not tree item is part of the selectable parent. */
    selectableParent?: boolean;

    indeterminate?: boolean;
  ```

  Reason for removal:

  1. These props are not in 0.47 version.
  2. `selected` and `indeterminate` can be accessed from the flattened tree. (In fact, if possible, I'd like to remove all props that can be accessed from the flattened tree, for example `parent`, `index`, ..., .etc). `selectableParent` is useless in my opinion.

- Tree does not create and pass `contentRef` prop to each TreeItem anymore.
Originally Tree creates and keeps a map of refs to each tree items, and assigned it to tree item's `contentRef` prop. The new re-worked tree does not do this anymore. 
But `contentRef` is still kept as a prop for TreeItem.

#### Focus areas to test

(optional)