Fixes custom error rendering in TextField (#16487)
#### Pull request checklist

- [x] Addresses an existing issue: Fixes #16049 
- [x] Include a change request file using `$ yarn change`

#### Description of changes

TextField renders all the error messages under a p tag with custom styles, that's fine as long as the error message is string.
When there is a custom error message (i.e., JSX.Element) rendering it under the p tag violates DOM validation (e.g, div can't be child of p). This PR addresses this issue.