MenuButton: Adding persistMenu prop to allow for constant rendering of the Menu (#16742)
#### Pull request checklist

- [ ] Addresses an existing issue: Fixes #0000
- [x] Include a change request file using `$ yarn change`

#### Description of changes

This PR adds the `persistMenu` prop to the `MenuButton` component that exists on the v7 version of the `Button`. This prop allows for constant rendering of the `Menu` independent if the `Menu` itself has been `expanded` or not. This is done by `hiding` the `Menu` instead of creating and destroying it, which could potentially improve the performance of the menu opening but impact the overall performance by having more elements in the DOM.