Fix publishing of webpack bundles and docs (#17246)
#### Pull request checklist

- [x] Addresses an existing issue: Fixes #17208, fixes #17213

#### Description of changes

The webpack 5 change #16447 broke bundles in production builds because the webpack CLI arg for turning on production mode has changed from `--production` to `--mode=production`. This was causing a build error, but it didn't cause the release build to fail because `scripts/runPublished.js` wasn't appropriately handling the exit code.

This PR fixes both issues, and switches the publish build/bundle to run using lage rather than lerna for consistency.