fix(Dropdown): prevent scrollable parent scroll when selecting an item (#17222)
#### Pull request checklist

- [ ] Addresses an existing issue: Fixes #0000
- [ ] Include a change request file using `$ yarn change`

#### Description of changes

In `Dropdown` we uses `Downshift`'s `scrollIntoView` to make sure the highlighted item is always in the view when the user opens dropdown.
Inside `scrollIntoView`, we use `computeScrollIntoView` to know how much we should scroll. However, if there's no `boundary` passed into `computeScrollIntoView`, all the parent elements of the target is checked, and sometimes it could scroll the viewport. 

For example:
![before](https://user-images.githubusercontent.com/28751745/109544259-a1689b80-7ac7-11eb-9a16-0c3387a2398a.gif)
The code can be found in this [codesandbox](https://codesandbox.io/s/ecstatic-lake-umf9t?file=/src/App.js)
Note: `positionFixed: true` on dropdown popper is important for this behavior to appear.

This PR adds a boundary to `computeScrollIntoView`, to make sure only the dropdown menu is checked for scrolling.
Behavior after fix:
![after](https://user-images.githubusercontent.com/28751745/109544336-b7765c00-7ac7-11eb-9a87-987206676514.gif)


#### Focus areas to test

(optional)