BREAKING CHANGE(MenuItem): use chevron as submenu indicator bg image (#17664)
### Description of changes

MenuItem uses a triangle background image for submenu indicator.
This PR changes this triangle to chevron, to be aligned with toolbarMenuItem's indicator.
>NOTE: it is still a background image!!!

RTL example in docsite is updated with submenu to reflect indicator change.

### Visual difference:

- before:
![image](https://user-images.githubusercontent.com/28751745/113283841-19a1c700-92e9-11eb-8db4-0f9adaf7e1ac.png)

- after:
![image](https://user-images.githubusercontent.com/28751745/113309049-0e5d9400-9307-11eb-80e2-edc614551887.png)

### Initiative behind this PR
TMP is trying to use _chevron_ for indicator. And it was difficult for them to get rid of the existing _triangle_ indicator in MenuItem.

My original plan was to switch to icon instead of background image, because I thought background image cannot be overriden. But it was my misunderstanding, it actually can be. 
In such case, the only advantage of changing background image to icon, would be easier styling of colors. But since TMP should be using fluent theme instead of custom color, this advantage is not so important.

Still, the overriding of background image is not intuitive, and TMP is not using triangle. So after a discussion with the team, we decide to change triangle to chevron so TMP does not need to override it. And also it is aligned with the indicator of ToolbarMenuItem.